jquery-fallr plugin and other major changes.

+++ public/javascripts/boards.js
@@ -36,55 +36,54 @@ var BoardCollection = Backbone.Collection.extend({ model: Board })
  * Views
  */
 
-// Actually this is the boards window (not an item)
 var BoardsView = Backbone.View.extend({
 
-  className: "boards_window",
+  className: "boards-window",
 
   template: JST["boards/boards"],
 
   events: {
-    "click .ui-overlay": "close",
-    "click .open_board": "openBoard",
     "click .new_board": "createBoard",
-    "click .destroy_board": "removeBoard"
+    "click .destroy_board": "removeBoard",
+    "click .boards-items li": "selectItem"
   },
 
   initialize: function() {
-    _.bindAll(this, "render", "close", "escKey", "openBoard", "createBoard", "removeBoard");
-
-    j(document).bind("keyup", this.escKey);
+    _.bindAll(this, "render", "openBoard", "createBoard", "removeBoard", "selectItem");
   },
 
-  escKey: function(event) {
-    if (event.keyCode == 27) {
-      this.close();
+  selectItem: function(event) {
+    j(event.currentTarget).siblings().removeClass("selected");
+    j(event.currentTarget).addClass("selected");
+  },
+
+  openBoard: function() {
+    var selected = j(".boards-items").find(".selected"), that = this;
+
+    if (selected.length) {
+      var id = selected.data("id");
+
+      currentBoard = Boards.get(id);
+
+      Stickies.fetch({
+        success: function(collection, response) {
+          j(".stickies").empty();
+          j(".board_title").text(currentBoard.get("title"));
+
+          collection.trigger("reset");
+
+          that.close();
+        }
+      });
+
+      Quadro.views.workspaceView.shareMenuView.render(true);
     }
   },
 
-  openBoard: function(event) {
-    var id = j(event.currentTarget).parents("li").data("id"), that = this;
-
-    currentBoard = Boards.get(id);
-
-    Stickies.fetch({
-      success: function(collection, response) {
-        j(".stickies").empty();
-        j(".board_title").text(currentBoard.get("title"));
-
-        collection.trigger("reset");
-
-        that.close();
-      }
-    });
-
-    Quadro.views.workspaceView.shareMenuView.render(true);
-
-    event.preventDefault();   
-  },
-
   createBoard: function(event) {
-    var title = prompt("Enter a title:") || "Untitled";
+    var title = prompt("Enter a title:");
+
+    if (title == "" || title == null) return;
 
     Boards.create({ title: title }, { 
       success: function(model, response, xhr) {
@@ -92,42 +91,60 @@ var BoardsView = Backbone.View.extend({
 
         j(item)
           .css("display", "none")
-          .appendTo(".boards_window ul")
+          .appendTo(".boards-items")
           .slideDown();
       }
     });
-
-    event.preventDefault();
   },
 
-  removeBoard: function(event) {
-    var el = j(event.currentTarget)
-      , id = el.parent().data("id")
-      , entry = Boards.get(id)
-      , confirmed = confirm("Are you sure?");
+  removeBoard: function() {
+    var selected = j(".boards-items").find(".selected");
 
-    // Finally, destroy entry on database
-    if (confirmed) {
-      entry.destroy({
-        success: function(model, response) {
-          el.parent().slideUp();
-        },
-        error: function(model, xhr, callbacks) {
-          var response = JSON.parse(xhr.responseText);
-          alert(response.message);
-        }
-      });
+    if (selected.length) {
+      var id = selected.data("id")
+        , entry = Boards.get(id)
+        , confirmed = confirm("Are you sure?");
+
+      // Finally, destroy entry on database
+      if (confirmed) {
+        entry.destroy({
+          success: function(model, response) {
+            selected.slideUp();
+          },
+          error: function(model, xhr, callbacks) {
+            var response = JSON.parse(xhr.responseText);
+            alert(response.message);
+          }
+        });
+      }
     }
-
-    event.preventDefault();
   },
 
   open: function() {
-    j(this.el).show();
+    if (j("#fallr:visible").length) return;
+
+    j.fallr('show', {
+      zIndex: 999,
+      closeKey: true,
+      closeOverlay: true,
+      icon: 'basket',
+      width: '320px',
+      height: '280px',
+      position: 'center',
+      content: j(this.render().el).clone(true),
+      buttons: {
+        button1: { className: 'new-icon', text: 'New', onclick: this.createBoard },
+        button3: { className: 'remove-icon', text: 'Delete', onclick: this.removeBoard },
+        button2: { className: 'open-icon', text: 'Open', onclick: this.openBoard }
+      }
+    });
+
+    j("#fallr").css({ width: "265px", padding: "15px" });
+    j("#fallr-buttons").css("position", "absolute");
   },
 
   close: function() {
-    j(this.el).hide();
+    j.fallr('hide');
   },
 
   render: function() {
